Fixed: Incorrect Total PO Items, wrong Current Sales Order data and missing parent product name (149) (143) (141)
The Purchase Order card is displaying the wrong corresponding sales order data for the Purchase order. It is counting the total number of sales orders associated with the purchase order.

The Total PO items displayed in the Purchase Order card includes both Sales orders and Purchase orders, resulting in an incorrect net total.

The products on the Catalog page only display their variant names, which makes it difficult to identify them without their parent product names.